requirements.txt
# Python dependencies for lab/eval/ scoring framework
PyYAML>=6.0.3,<7.0
pytest>=9.0.3,<10.0
